Move redisassert.* to valkeyassert.*
