+++ src/ae.c
@@ -32,7 +32,7 @@
 
 #include "ae.h"
 #include "anet.h"
-#include "redisassert.h"
+#include "valkeyassert.h"
 
 #include <stdio.h>
 #include <sys/time.h>
