+++ src/intset.c
@@ -34,7 +34,7 @@
 #include "intset.h"
 #include "zmalloc.h"
 #include "endianconv.h"
-#include "redisassert.h"
+#include "valkeyassert.h"
 
 /* Note that these encodings are ordered, so:
  * INTSET_ENC_INT16 < INTSET_ENC_INT32 < INTSET_ENC_INT64. */
